+++ NEWS.md
@@ -1,5 +1,11 @@
 <!-- NEWS.md is maintained by https://cynkra.github.io/fledge, do not edit -->
 
+# tiltIndicator 0.0.0.9211
+
+* The `sector*()` functions now preserve unmatched products (@AnneSchoenauer
+@Tilmon #738). This changes the structure of the output at company level, which
+now has four levels of risk_category: "high", "low", "medium", and `NA`.
+
 # tiltIndicator 0.0.0.9209
 
 * The `emissions*()` functions now preserve unmatched products and missing